requirements.txt
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
docopt==0.6.2
idna==3.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
pathspec==0.9.0
pipreqs==0.4.11
platformdirs==2.4.1
protobuf==3.18.1
pyTelegramBotAPI==4.1.1
requests==2.26.0
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
yarg==0.1.9